Moving FTR charts to us_nd in program_evaluation
