+++ src/components/charts/program_evaluation/us_nd/free_through_recovery/FtrReferralsByGender.js
@@ -18,10 +18,10 @@
 import React, { useState, useEffect } from 'react';
 
 import { Bar } from 'react-chartjs-2';
-import { COLORS_STACKED_TWO_VALUES_ALT, COLORS } from '../../../assets/scripts/constants/colors';
-import { sortByLabel } from '../../../utils/dataOrganizing';
-import { configureDownloadButtons } from '../../../assets/scripts/utils/downloads';
-import { genderValueToHumanReadable, toInt } from '../../../utils/variableConversion';
+import { COLORS_STACKED_TWO_VALUES_ALT, COLORS } from '../../../../../assets/scripts/constants/colors';
+import { sortByLabel } from '../../../../../utils/dataOrganizing';
+import { configureDownloadButtons } from '../../../../../assets/scripts/utils/downloads';
+import { genderValueToHumanReadable, toInt } from '../../../../../utils/variableConversion';
 
 const FtrReferralsByGender = (props) => {
   const [chartLabels, setChartLabels] = useState([]);
