+++ src/components/charts/program_evaluation/us_nd/free_through_recovery/FtrReferralsByLsir.js
@@ -18,9 +18,9 @@
 import React, { useState, useEffect } from 'react';
 
 import { Bar } from 'react-chartjs-2';
-import { COLORS } from '../../../assets/scripts/constants/colors';
-import { configureDownloadButtons } from '../../../assets/scripts/utils/downloads';
-import { toInt } from '../../../utils/variableConversion';
+import { COLORS } from '../../../../../assets/scripts/constants/colors';
+import { configureDownloadButtons } from '../../../../../assets/scripts/utils/downloads';
+import { toInt } from '../../../../../utils/variableConversion';
 
 const FtrReferralsByLsir = (props) => {
   const [chartLabels, setChartLabels] = useState([]);
